requirements.txt
streamlit==1.54.0
pdfplumber==0.11.9
python-dotenv==1.2.1
requests==2.32.5
langchain-text-splitters==1.1.1
supabase==2.28.0
lightrag-hku==1.4.9.11
tiktoken==0.7.0
numpy==1.26.4
